backend/requirements.txt
Flask
Flask-Cors
numpy
Pillow
werkzeug
opencv-python
mongoengine
flask-bcrypt
PyJWT
python-dotenv
